requirements.txt
fastapi[all]
psycopg2==2.9.9
requests==2.31.0
SQLAlchemy==2.0.29
alembic==1.13.1
sqladmin==0.16.1
python-jose==3.3.0
cryptography==42.0.5
passlib==1.7.4
bcrypt==4.1.2